% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect}
\alias{connect}
\title{Connect to EMS and get the Auth token}
\usage{
connect(usr, pwd, proxies = NULL)
}
\arguments{
\item{usr}{String, FOQA username}

\item{pwd}{String, FOQA password}

\item{proxies}{List containing the system proxy information. The list contains attributes "url", "port", "usr", "pwd"}
}
\value{
a Connection object.
}
\description{
Connect to EMS and get the Auth token
}

